requirements.txt
numpy
pillow
flask
tensorflow
